feat: export functions to make custom matchers
To allow creating custom matchers, export all the necessary functions
from the `/utils` path.
